requirements.txt
absl-py==1.1.0
accelerate==0.11.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
blessed==1.20.0
blobfile==2.1.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
contextlib2==21.6.0
cycler==0.11.0
Deprecated==1.2.13
docker-pycreds==0.4.0
einops==0.4.1
einops-exts==0.0.3
elasticdeform==0.5.1
ema-pytorch==0.0.8
filelock==3.15.4
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google-auth==2.9.0
google-auth-oauthlib==0.4.6
gpustat==1.1.1
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.17.3
humanize==4.2.2
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
Jinja2==3.1.4
joblib==1.1.0
kiwisolver==1.4.3
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.5
matplotlib==3.5.2
ml_collections==0.1.1
monai==0.9.0
mpmath==1.3.0
multidict==6.0.2
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.10.0.84
packaging==24.1
pandas==1.4.3
pathtools==0.1.2
Pillow==9.1.1
promise==2.3
protobuf==3.20.1
psutil==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodomex==3.21.0
pyDeprecate==0.3.1
pydicom==2.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.4
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
#pyzmq==19.0.2
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.1.5
rsa==4.8
safetensors==0.4.4
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1.2
six==1.16.0
sk-video==1.1.10
smmap==5.0.0
sympy==1.13.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
threadpoolctl==3.1.0
tifffile==2022.8.3
timm==1.0.8
toml==0.10.2
torch==2.1.2
torch-tb-profiler==0.4.0
torchio==0.18.80
torchmetrics==0.9.1
torchstat==0.0.7
torchvision==0.16.2
tqdm==4.64.0
triton==2.1.0
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.21
wcwidth==0.2.13
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0